Add twitter support
Twitter support requires a consumer key/secret as well as an access
token/secret. The consumer key/secret are generated by twitter when you
sign up for an account. The access token/secret can be generated using
the twitter_auth.rb script and authorising the application as a logged
in twitter user.

Also some refactoring/tidy ups of a few pieces.